Fix for warning under Windows.
git-svn-id: http://www.openscenegraph.org/svn/osg/OpenThreads/trunk@2253 16af8721-9629-0410-8352-f15c8da7e697

+++ win32_src/Win32Thread.cpp
@@ -379,7 +379,7 @@ int Thread::testCancel()
 
     DWORD curr = GetCurrentThreadId();
 
-    if( pd->uniqueId != curr )
+    if( pd->uniqueId != (int)curr )
         return -1;
 
     pd->isRunning = false;
